Gastepo/requirements.txt
allure_python_commons==2.8.6
allure_pytest==2.8.6
pymongo==3.9.0
xlrd==1.2.0
colorlog==4.0.2
openpyxl==3.0.0
xlutils==2.0.0
requests==2.22.0
SQLAlchemy==1.3.9
jsonpath==0.82
xlwt==1.3.0
rsa==4.0
pandas==0.25.3
retrying==1.3.3
xmltodict==0.12.0
pytest==5.3.2
PyMySQL==0.9.3
redis==3.3.10
Flask==1.1.1
pycryptodome==3.9.7
PyHamcrest==2.0.2
PyYAML==5.3.1
fastapi==0.58.1
uvicorn==0.11.5
aiofiles==0.5.0
jenkins==1.0.2